can_backend/custom_permission.py
from rest_framework import permissions
from rest_framework.response import Response
from rest_framework import status

class CustomIsAuthenticated(permissions.BasePermission):
    def has_permission(self, request, view):
        if request.user.is_authenticated:
            return True
        return Response({"detail": "Authentication credentials were not provided."}, status=status.HTTP_401_UNAUTHORIZED)