requirement.txt
anyio==4.2.0
asgiref==3.7.2
cachetools==5.3.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.2
defusedxml==0.7.1
Django==5.0.1
django-allauth==0.61.1
django-environ==0.11.2
djangorestframework==3.14.0
google-auth==2.27.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
oauthlib==3.2.2
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
sniffio==1.3.0
sqlparse==0.4.4
urllib3==2.2.0
